Weekly Commit. Basic Decision Tree with Flee Behavior completed. Minor bug regarding the enemy sprite instantly animating to the correct angle that needs to be fixed.

+++ SimpleShooterV1/PatrolBehavior.cpp
@@ -3,19 +3,19 @@
 
 Vector2 PatrolBehavior::Wander(Vector2 entityPos, Vector2 targetPos)
 {
-    Vector2 wanderTarget = targetPos + Vector2(RandomClamped(1, -1) * patrolJitter, RandomClamped(1, -1) * patrolJitter);
+    Vector2 wanderTarget = targetPos + Vector2((float) (RandomClamped(1, -1) * patrolJitter), (float) (RandomClamped(1, -1) * patrolJitter));
     wanderTarget = wanderTarget.Normalized();
 
 
     //Scale with circle radius
-    wanderTarget = wanderTarget * patrolRadius;
+    wanderTarget = wanderTarget * (float) patrolRadius;
 
 
-    Vector2 circleCenter = entityPos.Normalized() * patrolDistance;
-    Vector2 displacement = Vector2(0, -1) * patrolRadius;
+    Vector2 circleCenter = entityPos.Normalized() * (float) patrolDistance;
+    Vector2 displacement = Vector2(0, -1) * (float) patrolRadius;
 
-    displacement.x *= patrolJitter;
-    displacement.y *= patrolJitter;
+    displacement.x *= (float) patrolJitter;
+    displacement.y *= (float) patrolJitter;
 
 
     return circleCenter + displacement;
